Adding several validations for primitives in Wagic
This code appears to be checking if a particular line starts with the string "abilities=". If it does, it then checks if the line contains any of a list of keywords by iterating over the list of keywords and checking if each keyword is contained within the line. If the line does not contain any of the keywords, it prints a message to the console indicating that fact and includes the line number and the contents of the line.